giskard-client/pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
giskard = "giskard.cli:cli"

[tool.poetry]
name = "giskard"
version = "1.7.0"
description = "Inspect your AI models visually, find bugs, give feedback 🕵️‍♀️ 💬"
readme = "README.md"
authors = ["Giskard AI <hello@giskard.ai>"]
license = "Apache Software License 2.0"
repository = "https://github.com/Giskard-AI/giskard-client"
homepage = "https://github.com/Giskard-AI/giskard-client"
include = ["giskard/ml_worker/generated/*"]

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = [
    "Artificial Intelligence",
    "Machine Learning",
    "Quality",
    "MLOps"
]

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]



[tool.poetry.dependencies]
python = ">=3.7.5,<3.11"
cloudpickle = "^2.1.0"
importlib_metadata = { version = "^4.11.4" }
numpy = ">=1.21.6,<1.22.0"
pandas = "^1.3.5"
pydantic = "^1.10.2"
requests-toolbelt = "^0.9.1"
scikit-learn = ">=1.0.0,<1.1.0"
mixpanel = "^4.10.0"
beautifulsoup4 = "^4.11.1"
eli5 = "^0.13.0"
grpcio = "^1.46.3"
grpcio-status = "^1.46.3"
protobuf = "^3.9.2"
tenacity = "^8.1.0"
python-daemon = "^2.3.1"
click = "^8.1.3"
lockfile = "^0.12.2"
psutil = "^5.9.2"
requests = "^2.28.1"
tqdm = "^4.64.1"
setuptools = "^65.4.1"
zstandard = "^0.15.2"
scipy = ">=1.7.2,<1.8"
shap = "^0.41.0"
ipython = "^7.0.0"

[tool.poetry.group.dev.dependencies]
typing-extensions = "^4.1.0"
black = "^22.8.0"
bandit = "^1.7.4"
darglint = "^1.8.1"
flake8 = "^3.9.2"
httpretty = "^1.1.4"
isort = { extras = ["colors"], version = "^5.10.1" }
jupyter = "^1.0.0"
jupyterlab = "^3.4.2"
pre-commit = "^2.19.0"
pydocstyle = "^6.1.1"
pylint = "^2.13.9"
pyupgrade = "^2.32.1"
safety = "^1.10.3"
grpcio-tools = "^1.46.3"
mypy-protobuf = "^3.2.0"
mypy-extensions = "^0.4.3"
mypy = "^0.982"
deptry = "^0.5.13"

[tool.poetry.group.test.dependencies]
pytest-cov = "^3.0.0"
pytest = "^7.1.2"
catboost = "^1.1"

[tool.black]
# https://github.com/psf/black
target-version = ["py37"]
line-length = 100
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 37
line_length = 100

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.7
pretty = true
show_traceback = true
color_output = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
plugins = ["pydantic.mypy"]

allow_redefinition = false
check_untyped_defs = false # TODO: Andrey: fix it gradually
disallow_untyped_defs = false # TODO: Andrey: fix it gradually
disallow_any_generics = false # TODO: Andrey: fix it gradually
disallow_incomplete_defs = false # TODO: Andrey: fix it gradually
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = false # TODO: Andrey: fix it gradually
strict_equality = false # TODO: Andrey: fix it gradually
strict_optional = false # TODO: Andrey: fix it gradually
warn_no_return = false
warn_redundant_casts = true
warn_return_any = false
warn_unreachable = true
warn_unused_configs = false
warn_unused_ignores = false


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs = ["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
# Extra options:
addopts = [
    "--strict-markers",
    "--tb=short",
    "--doctest-modules",
    "--doctest-continue-on-failure",
]

[virtualenvs]
create = true
in-project = true
